Expose testing utils to plugins
