+++ src/sqlfluff/testing/__init__.py
@@ -1,0 +1,1 @@
+"""Testing utils we want to expose for usage by plugins."""
