+++ test/core/rules/rule_test_cases_test.py
@@ -1,28 +1,13 @@
 """Runs the rule test cases."""
 import os
-from glob import glob
-
 import pytest
-import oyaml as yaml
+from sqlfluff.testing.rules import load_test_cases, rules__test_helper
 
-from .std_test import rules__test_helper, RuleTestCase
-
-ids = []
-test_cases = []
-
-test_cases_path = os.path.join(
-    os.path.abspath(os.path.dirname(__file__)), "test_cases", "*.yml"
+ids, test_cases = load_test_cases(
+    test_cases_path=os.path.join(
+        os.path.abspath(os.path.dirname(__file__)), "test_cases", "*.yml"
+    )
 )
-
-for path in sorted(glob(test_cases_path)):
-    with open(path) as f:
-        raw = f.read()
-
-    y = yaml.safe_load(raw)
-
-    rule = y.pop("rule")
-    ids.extend([rule + "_" + t for t in y])
-    test_cases.extend([RuleTestCase(rule=rule, **v) for k, v in y.items()])
 
 
 @pytest.mark.parametrize("test_case", test_cases, ids=ids)
